Reverted Additional Module/Package
