+++ src/databricks/labs/ucx/cli.py
@@ -13,7 +13,6 @@ from databricks.labs.ucx.hive_metastore import ExternalLocations, TablesCrawler
 from databricks.labs.ucx.hive_metastore.mapping import TableMapping
 from databricks.labs.ucx.install import WorkspaceInstaller
 from databricks.labs.ucx.installer import InstallationManager
-from databricks.labs.ucx.migration.migrate_tables import TablesMigrate
 
 logger = logging.getLogger("databricks.labs.ucx")
