+++ src/databricks/labs/ucx/hive_metastore/__init__.py
@@ -1,5 +1,6 @@
 from databricks.labs.ucx.hive_metastore.grants import GrantsCrawler
-from databricks.labs.ucx.hive_metastore.locations import ExternalLocations, Mounts
-from databricks.labs.ucx.hive_metastore.tables import TablesCrawler
+from databricks.labs.ucx.hive_metastore.locations import ExternalLocations, Mount, Mounts
+from databricks.labs.ucx.hive_metastore.tables import TablesCrawler, TablesMigrate
+from databricks.labs.ucx.hive_metastore.mapping import TableMapping
 
-__all__ = ["TablesCrawler", "GrantsCrawler", "Mounts", "ExternalLocations"]
+__all__ = ["TablesCrawler", "TablesMigrate", "GrantsCrawler", "ExternalLocations", "Mount", "Mounts", "TableMapping"]
