+++ src/databricks/labs/ucx/hive_metastore/mapping.py
@@ -11,8 +11,9 @@ from databricks.sdk.service.workspace import ImportFormat
 
 from databricks.labs.ucx.account import WorkspaceInfo
 from databricks.labs.ucx.framework.crawlers import StatementExecutionBackend
-from databricks.labs.ucx.hive_metastore import TablesCrawler
-from databricks.labs.ucx.hive_metastore.tables import Table
+
+
+from databricks.labs.ucx.hive_metastore.tables import Table, TablesCrawler
 
 logger = logging.getLogger(__name__)
